chore(exercises): update exercise progress

+++ exercises/02_functions/functions3.rs
@@ -1,4 +1,4 @@
-fn call_me(num: u8) {
+fn call_me(num: i32) {
     for i in 0..num {
         println!("Ring! Call number {}", i + 1);
     }
